Rename paths.lua to pathfinding.lua
